Coerce inputs for method calls and find available operators for an object
I am adding the ability for inputs into methods to be have their value coerced incase you are just allowing users to type in values via a UI.  I am also starting on adding the ability to inspect an object and find all of the possible MemberNames and Operators that can be executed on those.   This part is still a little rough.